UpdateMoveDirection & UpdateJump changes in CombinedPlayerController
I made two changes:

In the UpdateMoveDirection() method, I added the vertical input axis to the relative direction calculation. This should make the character move at the correct speed in any direction.

In the UpdateJump() method, I moved the remainingJumps = 2; line to the beginning of the method to reset the remaining jumps when the character is grounded. This should enable double jumping.