Movement changes / World Space 4 Update - Working State
BoostPad.cs:

Added logic to apply a boost force to the player based on the object's local X-axis instead of the world's X-axis.
BoostPad_Y_Axis.cs:

Created a new script to apply a boost force to the player based on the object's local Y-axis, stopping the player's X and Z momentum.
ColliderManager.cs:

Created a new script that enables and disables colliders based on their distance from the player, except for those in the "alwaysEnabledColliders" list.
ShadowManager.cs:

Created a new script that disables the shadows of objects in the "shadowCastingObjectsToDisable" list if they are at a certain distance from the player. The shadows of objects not in the list will render normally.
CustomCulling.cs:

Created a new script that implements custom culling by disabling the MeshRenderer of objects that are not in the camera's view frustum. The script supports adjustable culling distances and updates the culling status based on the player's current position.
CombinedBoost.cs:

Updated the script to have a Transform named "boostDirectionReference" to determine the boost direction.
Fixed the audio issue by ensuring the AudioSource component is properly assigned to the audioSource variable.
CombinedPlayerController.cs:

Added functionality to reset the jump count after the player uses a boost, allowing them to jump after using a boost.