WorldSpace 4 + 5 Updates / MusicManger Weights / WorldSpace 6 Created / Updated Collectible Count
Added weights to the likelihood of particular tracks playing:

Created a new float[] named trackWeights to store the weights for each track.
Modified the PlayShuffledMusic() method to choose tracks based on their weights.

DoorSwing Script:
Added tooltips that will help explain the functionality of the script and provide guidance on how to use it in the Unity Editor.