Interactive trigger update.
It includes the locked SFX array, cooldown for playing SFX, and the cubes remaining UI. When the player tries to open the locked door, it will play the standalone locked SFX as well as a random sound from the array, and there will be a cooldown so the SFX cannot be spammed. The cubes remaining UI will be shown when the player is in range and will update accordingly.