GameManager script / Player FPS zone script update. dungeon 1 update / main hub updated and organized /created the CameraControlZone script.
updated version of the GameManager script replaces the numberOfWorldSpaces variable with a GameObject array called worldSpaces, which will store the WorldSpace parent objects. You can then assign these parent objects in the Unity Inspector.

In the Awake() method, the script now initializes the dungeonCompletionStatus array using the length of the worldSpaces array and sets all WorldSpaces, except the first one, to inactive. The UnlockNextWorldSpace() method has also been updated to activate the next WorldSpace by setting the corresponding game object to active.

I've added a new header "FPS Zone Assets" and a new public array of GameObjects called "fpsZoneAssets" to store the assets I want to enable/disable. A new private boolean variable "assetsEnabled" is added to keep track of whether the assets are enabled or not.

The ToggleFPSZoneAssets() method has been added to enable or disable the assets depending on the enable parameter.

The ToggleFirstPersonMode() coroutine now calls ToggleFPSZoneAssets() to enable or disable the assets when the camera mode changes.