added OutlineWhenOccluded script / modded FPSZoneTrigger script.
OutlineWhenOccluded. This script checks if the player's character is occluded by another object using a Raycast from the main camera's position to the player's position. If the player is occluded, it will enable the outline by adding the outline material to the player's renderer.

FPSZoneTrigger. To make sure the assets are enabled during the camera switch from FPS to Iso, I used a Coroutine to delay the enabling and disabling of assets. This way, I synchronized the camera switch and asset activation.

--

In the past few hours, I have made several updates and changes to my scripts to improve their functionality and achieve the desired features.

First, I updated the Outline script by adding a CheckVisibility() function to determine if the player is partially or completely hidden. This allowed me to control the outline behavior based on the player's visibility. I also modified the Update() method to call the CheckVisibility() function and apply the outline only when the player is hidden.

Next, I extended the Outline script functionality to outline nearby platforms when the player is hidden. I added new variables to control the color and range of outlined platforms, as well as a new method, GetNearbyPlatforms(), to find and outline the platforms within the specified range. I also modified the CheckVisibility() function to apply the outline to the nearby platforms.

Finally, I updated the FPSZoneTrigger script to synchronize the camera switch and asset activation by using Coroutines. I modified the OnTriggerExit() method to call StartCoroutine() and added a new Coroutine method, ToggleAssetsWithDelay(), to delay the enabling and disabling of assets. This ensures that the assets are activated during the camera switch, preventing them from being visible during the transition.