updated Pillar Pickup and Unlock scripts / logic
added two lists to the PillarManager script, one for the assets to be disabled and another for the assets to be enabled.

added new public fields for the SFX and VFX prefabs, and a field for the AudioSource component. Then, modified the HoldObject() and DropObject() methods to play the SFX, reset the rotation, and instantiate the VFX.

I added a new variable called "activationSound" to the RunePressurePlate script to hold the activation sound. In the SetActivated method, I added a conditional statement to check if the pressure plate has been activated, and if so, play the activation sound using AudioSource.PlayClipAtPoint. The sound will play at the position of the pressure plate.