started work on objective scripts.
created two new scripts, LockedDoor and DoorKeyUnlock. The LockedDoor script will handle the logic for the locked door, displaying the UI prompt, and playing the locked SFX. The DoorKeyUnlock script will handle the logic for the key and unlocking the various assets.

modified version of the LeverController script that includes the "Find a way to turn on the power" prompt and only activates when the power is turned on.

--
To future Ray:

To set up the scripts properly, follow these steps:

Create and attach the scripts to their respective objects:
Attach the LockedDoor script to each door object that needs to be locked.
Attach the DoorKeyUnlock script to the key object that unlocks the doors.
Attach the modified LeverController script to each lever object.
Set up the LockedDoor script for each locked door:
In the Inspector, assign the locked SFX AudioClip to the lockedSfx field.
Assign a UI prompt text (TextMeshPro) object to the objectiveUI field.
Set the prompt text for when the player approaches the locked door.
Set up the DoorKeyUnlock script for the key object:
Create an array of locked doors in the doorsToUnlock field and assign the LockedDoor script component from each locked door to this array.
Create an array of levers that need power activation in the leversToActivatePower field and assign the LeverController script component from each lever to this array.
Set up the LeverController script for each lever:
Assign appropriate UI elements and assets to their respective fields in the Inspector as described in the original LeverController script.
Assign a UI prompt text (TextMeshPro) object to the powerActivationUI field.
Set the prompt text for when the player approaches the lever without the power activated.
Set up the colliders and parent objects:
For each locked door, create a parent object with a trigger collider larger than the door objects themselves. Set the isTrigger property of the colliders to true.
For the key object, create a collider and set the isTrigger property to true.
For each lever object, make sure there is a collider with the isTrigger property set to true.
Set up object tags:
Ensure that the player object is tagged with the "Player" tag so that trigger events can identify the player correctly.